+++ src/libponyrt/pony.h
@@ -205,6 +205,28 @@ void pony_gc_send(pony_ctx_t* ctx);
  */
 void pony_gc_recv(pony_ctx_t* ctx);
 
+/** Start gc tracing for acquiring.
+ * 
+ * Call this when acquiring objects. Then trace the objects you want to
+ * acquire, then call pony_acquire_done. Acquired objects will not be GCed
+ * until released even if they are not reachable from Pony code anymore.
+ * Acquiring an object will also acquire all objects reachable from it as well
+ * as their respective owners. When adding or removing objects from an acquired
+ * object graph, you must acquire anything added and release anything removed.
+ * A given object (excluding actors) cannot be acquired more than once in a
+ * single pony_gc_acquire/pony_acquire_done round. The same restriction applies
+ * to release functions.
+ */
+void pony_gc_acquire(pony_ctx_t* ctx);
+
+/** Start gc tracing for releasing.
+ * 
+ * Call this when releasing acquired objects. Then trace the objects you want
+ * to release, then call pony_release_done. If an object was acquired multiple
+ * times, it must be released as many times before being GCed.
+ */
+void pony_gc_release(pony_ctx_t* ctx);
+
 /** Finish gc tracing for sending.
  *
  * Call this after tracing the GCable contents.
@@ -216,6 +238,18 @@ void pony_send_done(pony_ctx_t* ctx);
  * Call this after tracing the GCable contents.
  */
 void pony_recv_done(pony_ctx_t* ctx);
+
+/** Finish gc tracing for acquiring.
+ * 
+ * Call this after tracing objects you want to acquire.
+ */
+void pony_acquire_done(pony_ctx_t* ctx);
+
+/** Finish gc tracing for releasing.
+ * 
+ * Call this after tracing objects you want to release.
+ */
+void pony_release_done(pony_ctx_t* ctx);
 
 /** Trace memory
  *