+++ CHANGELOG.md
@@ -61,6 +61,7 @@ All notable changes to the Pony compiler and standard library will be documented
 - ponytest TestHelper.dispose_when_done
 - copysign and infinite for floating point numbers
 - contains() method on Array
+- GC tracing with acquire/release semantics.
 
 ### Changed